🐛 fix: 홈화면 main 컴포넌트 페이지로 이동
